Implement service discovery, agent
Again inspired heavily by Distributed Services by Go, this adds service
discovery with Serf/SWIM with an agent which is repsonsible for
configuring the server, service discovery, Raft etc.

Server now takes a store interface, allowing us to swap in a
DistributedStore instead of a Store.

cmux is used to multiplex Raft and gRPC on the same port.