Implement key deletion
Deleted keys are tombstoned with a nil value. I'm not sure if this is
really the best approach, since it should theoretically be possible for
someone to set a key to an empty value.

An alternative would be to add a "deleted" bool flag to KeyValue in
sstable and have a separate map of deleted keys in storage.Store.

With no support for levelled compaction, tombstones are currently
retained forever.