Add server.controller
Controller is responsible for holding the writable memory store and all
SSTables associated with a server. It will write values to the memtable
until a threshold number of bytes is written, then flush the values to a
new SSTable.

Values from newer SSTables take precedence over older tables. Values
stored in the memtable will be lost if the controller is not closed
cleanly. There is as yet no WAL (except that which Raft provides, but
it's not clear to me yet how Raft's logs would interact with
non-volatile SSTables.

It may be that all writes require an incrementing ID (a log offset?) so
that we know from where to replay the log.