Controller now generates a transaction ID
It's not really clear to me how this interacts with Raft, or when the
cluster ends up with a netsplit. Is it possible for a new leader to
generate an earlier transaction ID? Do the followers need to keep track
of the most recently seen ID? What if the Raft log is replayed with IDs
which have already been seen?